从gitee同步

+++ unimall-app-api/src/main/java/com/iotechn/unimall/app/api/cart/CartServiceImpl.java
@@ -2,15 +2,12 @@ package com.iotechn.unimall.app.api.cart;
 
 import com.baomidou.mybatisplus.mapper.EntityWrapper;
 import com.iotechn.unimall.app.api.category.CategoryService;
-import com.iotechn.unimall.app.exception.AppExceptionDefinition;
-import com.iotechn.unimall.app.exception.AppServiceException;
+import com.iotechn.unimall.core.exception.ExceptionDefinition;
+import com.iotechn.unimall.core.exception.AppServiceException;
 import com.iotechn.unimall.core.exception.ServiceException;
-import com.iotechn.unimall.core.exception.ServiceExceptionDefinition;
 import com.iotechn.unimall.data.domain.CartDO;
 import com.iotechn.unimall.data.dto.CartDTO;
-import com.iotechn.unimall.data.dto.UserDTO;
 import com.iotechn.unimall.data.mapper.CartMapper;
-import com.iotechn.unimall.data.util.SessionUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -34,7 +31,7 @@ public class CartServiceImpl implements CartService {
     private CategoryService categoryService;
 
     @Override
-    @Transactional
+    @Transactional(rollbackFor = Exception.class)
     public Boolean addCartItem(Long skuId, Integer num, Long userId) throws ServiceException {
         List<CartDO> cartDOS = cartMapper.selectList(
                 new EntityWrapper<CartDO>()
@@ -60,7 +57,7 @@ public class CartServiceImpl implements CartService {
     }
 
     @Override
-    @Transactional
+    @Transactional(rollbackFor = Exception.class)
     public Boolean subCartItem(Long skuId, Integer num, Long userId) throws ServiceException {
         List<CartDO> cartDOS = cartMapper.selectList(
                 new EntityWrapper<CartDO>()
@@ -82,7 +79,7 @@ public class CartServiceImpl implements CartService {
     }
 
     @Override
-    @Transactional
+    @Transactional(rollbackFor = Exception.class)
     public Boolean removeCartItem(Long cartId, Long userId) throws ServiceException {
         return cartMapper.delete(
                 new EntityWrapper<CartDO>()
@@ -91,14 +88,14 @@ public class CartServiceImpl implements CartService {
     }
 
     @Override
-    @Transactional
+    @Transactional(rollbackFor = Exception.class)
     public Boolean removeCartItemBatch(String cartIdList, Long userId) throws ServiceException {
         if (StringUtils.isEmpty(cartIdList)) {
-            throw new AppServiceException(AppExceptionDefinition.APP_PARAM_CHECK_FAILED);
+            throw new AppServiceException(ExceptionDefinition.PARAM_CHECK_FAILED);
         }
         String[] split = cartIdList.split(",");
         if (split.length == 0) {
-            throw new AppServiceException(AppExceptionDefinition.APP_PARAM_CHECK_FAILED);
+            throw new AppServiceException(ExceptionDefinition.PARAM_CHECK_FAILED);
         }
         List<Long> array = new ArrayList<>(split.length);
         for (String idRaw : split) {
@@ -107,6 +104,14 @@ public class CartServiceImpl implements CartService {
         return cartMapper.delete(
                 new EntityWrapper<CartDO>()
                         .in("id", array)
+                        .eq("user_id", userId)) > 0;
+    }
+
+    @Override
+    @Transactional
+    public Boolean removeCartAll(Long userId) throws ServiceException {
+        return cartMapper.delete(
+                new EntityWrapper<CartDO>()
                         .eq("user_id", userId)) > 0;
     }
 
@@ -121,7 +126,7 @@ public class CartServiceImpl implements CartService {
         if (update > 0) {
             return num;
         }
-        throw new AppServiceException(AppExceptionDefinition.CART_UPDATE_FAILED);
+        throw new AppServiceException(ExceptionDefinition.CART_UPDATE_FAILED);
     }
 
     @Override
